Adds support user/machine-specific config via CMDER_CONFIG env
Launcher now looks for CMDER_CONFIG environment variable at startup and uses it for setting up a new config location.
If there is no configuration for particular tool (ConEmu/Clink/etc) the it falls back to default location for this tool (usually CMDER_ROOT/config)
Note that clink profile should be placed into `clink` subfolder if CMDER_CONFIG is defined by user. This is necessary since there will be no chance to detect if it is not exists in user-defined config folder otherwise.

This also removes Cmder binary as per discussion with @MartiUK in bliker/cmder/pull/368